move flask code to api from acc, remove roboto since uneccessary
