+++ acc/templates/index.html
@@ -3,7 +3,6 @@
 
 <head>
   <meta charset='UTF-8'>
-  <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
   <style>
     .modal {
       display: none;
@@ -28,7 +27,7 @@
       background: #f1f3f5;
       -webkit-font-smoothing: antialiased;
       -moz-osx-font-smoothing: grayscale;
-      font-family: 'Roboto', sans-serif
+      font-family: sans-serif
     }
 
     body,
@@ -230,6 +229,7 @@
       </div>
     </div>
   </div>
+
   <!--
     handles the functionality when the user submits the form
     sends POST request to /api/form
